parsing refactroing
